Obstacle & debug log marshalling Part of issue #1
This provides the necessary infrastructure to handle passing received
Obstacle messages through to the managed layer, as well as some utility
stuff to make getting debug messages from the native layer a little
easier.

Obstacles are passed by-value through reverse PInvoke to the managed
side, which lets us rely on the GC to clean up for us there, and means
we don't need to worry about hanging onto obstacles on the native side
after they've been passed through. This all happens on a background
thread, and later (not implemented yet) the obstacles will get
picked up by the UI thread for rendering.

A downside of the PInvoke approach is we need to mirror any C/C++
structs we use with C# structs on the managed side. The main alternative
to avoid this is to use a C++/CLI wrapper between the C# and native C++
layers, but this doesn't seem to be supported by Unity at the moment.