Accumulate footstep angles from phi_leg_rel This should resolve issue #17. Each new footstep includes the relative angle from the previous step, so we should get the correct angle by adding all previous angles together to arrive at the angle for the current step.
On every new timestamp, only the angle of the first step from that
timestamp is kept, and all steps for the next timestamp are computed
relative to it (rather than relative to the last step of the previous
timestamp). This is because only the first step from each stamp was
actually taken, and the next real step *should* be relative to this,
rather than relative to the last planned-but-not-taken step (I think).